pyproject.toml
[project]
name = "nova-aetus"
version = "1.3.0"
description = "Institutional-grade Swing Trading System combining Technical, Fundamental, and Sentiment Analysis"
readme = "README.md"
requires-python = ">=3.12"
license = {text = "Proprietary"}
authors = [
    {name = "Nova Aetus Team"}
]
keywords = ["trading", "quantitative", "machine-learning", "xgboost", "swing-trading"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Financial and Insurance Industry",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Office/Business :: Financial :: Investment",
]

# Core dependencies (from requirements.txt)
dependencies = [
    "polars>=1.0.0",
    "pydantic>=2.0.0",
    "alpaca-py>=0.30.0",
    "yahooquery>=2.4.0",
    "xgboost>=2.0.0",
    "optuna>=3.0.0",
    "optuna-integration[xgboost]>=0.5.0",
    "scikit-learn>=1.3.0",
    "ollama>=0.1.0",
    "discord-webhook>=1.0.0",
    "asyncpg>=0.29.0",
    "psycopg2-binary>=2.9.0",
    "streamlit>=1.28.0",
    "plotly>=5.17.0",
    "python-dotenv>=1.0.0",
    "pytomlpp>=1.0.0",
    "mcp>=0.9.0",
    "typing-extensions>=4.8.0",
    "watchdog>=3.0.0",
    "aiohttp>=3.9.0",
]

# Platform-specific dependencies
[project.optional-dependencies]
# RAPIDS Memory Manager (Linux/Mac only)
rmm = ["rmm>=23.12.0; platform_system != 'Windows'"]

# Development dependencies
dev = [
    "pytest>=7.4.0",
    "pytest-asyncio>=0.21.0",
    "pytest-cov>=4.1.0",
    "pytest-mock>=3.11.0",
    "black>=23.0.0",
    "ruff>=0.1.0",
    "mypy>=1.5.0",
    "pre-commit>=3.0.0",
]

[project.urls]
Homepage = "https://github.com/your-org/nova-aetus"
Documentation = "https://github.com/your-org/nova-aetus/docs"
Repository = "https://github.com/your-org/nova-aetus"

[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

# Black code formatter configuration
[tool.black]
line-length = 100
target-version = ['py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  # directories
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | venv
  | _build
  | buck-out
  | build
  | dist
  | __pycache__
)/
'''

# Ruff linter configuration
[tool.ruff]
line-length = 100
target-version = "py312"
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
]

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]  # unused imports in __init__.py are OK
"tests/**/*.py" = ["ARG", "S101"]  # allow test function args and assert True

[tool.ruff.isort]
known-first-party = ["nova"]

# MyPy type checker configuration
[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false
disallow_incomplete_defs = false
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
show_error_codes = true

[[tool.mypy.overrides]]
module = [
    "alpaca.*",
    "yahooquery.*",
    "discord_webhook.*",
    "ollama.*",
]
ignore_missing_imports = true

# Pytest configuration (can migrate from pytest.ini)
[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
asyncio_mode = "auto"
addopts = [
    "-v",
    "--strict-markers",
    "--tb=short",
    "--cov=src/nova",
    "--cov-report=term-missing",
    "--cov-report=html",
]
markers = [
    "unit: Unit tests",
    "integration: Integration tests",
    "slow: Slow running tests",
    "benchmark: Performance benchmarks",
]
